fix workout_likes.post_workout_id

+++ db/migrate/20230604152941_create_workout_likes.rb
@@ -2,7 +2,7 @@ class CreateWorkoutLikes < ActiveRecord::Migration[6.1]
   def change
     create_table :workout_likes do |t|
       t.references :end_user, type: :bigint, foreign_key: true, null: false
-      t.references :post_workout, type: :integer, foreign_key: true, null: false
+      t.references :post_workout, type: :bigint, foreign_key: true, null: false
 
       t.timestamps
     end
